+++ src/Catalog/GeoFilter.php
@@ -1,0 +1,48 @@
+<?php
+/**
+ * @copyright Copyright (c) Wizacha
+ * @license Proprietary
+ */
+declare(strict_types=1);
+
+namespace Wizaplace\SDK\Catalog;
+
+final class GeoFilter
+{
+    /**
+     * @var float
+     */
+    private $latitude;
+
+    /**
+     * @var float
+     */
+    private $longitude;
+
+    /**
+     * @var int
+     */
+    private $radius;
+
+    public function __construct(float $latitude, float $longitude, int $radius)
+    {
+        $this->latitude = $latitude;
+        $this->longitude = $longitude;
+        $this->radius = $radius;
+    }
+
+    public function getLatitude(): float
+    {
+        return $this->latitude;
+    }
+
+    public function getLongitude(): float
+    {
+        return $this->longitude;
+    }
+
+    public function getRadius(): int
+    {
+        return $this->radius;
+    }
+}