+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 
 ### New features
 
+- Added an optional `$geoFilter` parameter to `\Wizaplace\SDK\Catalog\CatalogService::search`
+
 ### Bugfixes
 
 </details>